scripts/requirements-tools.txt
# Ubuntu 24+: use venv (run-process-logos.sh creates .venv-logo automatically):
#   bash scripts/run-process-logos.sh
# Or: .venv-logo/bin/pip install -r scripts/requirements-tools.txt
pillow>=10.0.0
numpy>=1.24.0
